refactor: app state
